requirements.txt
django
langchain
psycopg2
spacy
langchain
langchain_google_genai
numpy
pandas
scikit-learn
imblearn
transformers
torch
datasets
alt-profanity-check
huggingface_hub
bertopic
python-dateutil
google-auth
google-cloud-core
neo4j
markdown2